Added unit first sample test cases , and its prerequisites
