+++ src/main/java/com/dpoint/tutorial/Main.java
@@ -1,0 +1,14 @@
+package com.dpoint.tutorial;
+
+import org.camunda.bpm.spring.boot.starter.annotation.EnableProcessApplication;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+@EnableProcessApplication
+public class Main {
+    public static void main(String args[]) {
+        SpringApplication.run(Main.class);
+    }
+
+}
