+++ src/main/java/com/dpoint/tutorial/delegates/InstantiateProcessByMessageDelegate.java
@@ -1,0 +1,13 @@
+package com.dpoint.tutorial.delegates;
+
+import org.camunda.bpm.engine.RuntimeService;
+import org.camunda.bpm.engine.delegate.DelegateExecution;
+import org.camunda.bpm.engine.delegate.JavaDelegate;
+
+public class InstantiateProcessByMessageDelegate implements JavaDelegate {
+
+    public void execute(DelegateExecution execution) throws Exception {
+        RuntimeService runtimeService = execution.getProcessEngineServices().getRuntimeService();
+        runtimeService.startProcessInstanceByMessage("instantiationMessage");
+    }
+}
