+++ src/main/java/com/dpoint/tutorial/mocdel/CamundaCustomVariables.java
@@ -1,0 +1,14 @@
+package com.dpoint.tutorial.mocdel;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class CamundaCustomVariables<T extends Object> {
+    T variables;
+
+    public CamundaCustomVariables(T variables) {
+        this.variables = variables;
+    }
+}
