+++ src/main/java/com/dpoint/tutorial/mocdel/CamundaValueTypePairDto.java
@@ -1,0 +1,17 @@
+package com.dpoint.tutorial.mocdel;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class CamundaValueTypePairDto<T extends Object> {
+    T value;
+    String type;
+
+
+    public CamundaValueTypePairDto(T value) {
+        this.value = value;
+        this.type = value.getClass().getSimpleName();
+    }
+}
