+++ src/main/java/com/dpoint/tutorial/mocdel/NativeJsonDemoRequestDto.java
@@ -1,0 +1,13 @@
+package com.dpoint.tutorial.mocdel;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class NativeJsonDemoRequestDto {
+
+    String attribute1;
+    String attribute2;
+    String attribute3;
+}
