+++ src/main/java/com/dpoint/tutorial/mocdel/Summers.java
@@ -1,0 +1,26 @@
+package com.dpoint.tutorial.mocdel;
+
+public class Summers {
+    int num1;
+    int num2;
+    int num3;
+    int num4;
+
+    public int getNum1() {
+        return num1;
+    }
+
+    public int getNum2() {
+        return num2;
+    }
+
+    public int getNum3() {
+        return num3;
+    }
+
+    public int getNum4() {
+        return num4;
+    }
+
+
+}
