--- src/main/java/parentModule/childModule/Main.java
@@ -1,14 +1,0 @@
-package parentModule.childModule;
-
-import org.camunda.bpm.spring.boot.starter.annotation.EnableProcessApplication;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-@EnableProcessApplication
-public class Main {
-    public static void main(String args[]) {
-        SpringApplication.run(Main.class);
-    }
-
-}
