--- src/main/java/parentModule/childModule/mocdel/CamundaValueTypePairDto.java
@@ -1,17 +1,0 @@
-package parentModule.childModule.mocdel;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class CamundaValueTypePairDto<T extends Object> {
-    T value;
-    String type;
-
-
-    public CamundaValueTypePairDto(T value) {
-        this.value = value;
-        this.type = value.getClass().getSimpleName();
-    }
-}
