--- src/main/java/parentModule/childModule/mocdel/NativeJsonDemoRequestDto.java
@@ -1,13 +1,0 @@
-package parentModule.childModule.mocdel;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class NativeJsonDemoRequestDto {
-
-    String attribute1;
-    String attribute2;
-    String attribute3;
-}
