+++ src/main/resources/email-trials.bpmn
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<bpmn:definitions xmlns:bpmn="http://www.omg.org/spec/BPMN/20100524/MODEL" xmlns:bpmndi="http://www.omg.org/spec/BPMN/20100524/DI" xmlns:dc="http://www.omg.org/spec/DD/20100524/DC" xmlns:camunda="http://camunda.org/schema/1.0/bpmn" xmlns:di="http://www.omg.org/spec/DD/20100524/DI" id="Definitions_0e0r9n9" targetNamespace="http://bpmn.io/schema/bpmn" exporter="Camunda Modeler" exporterVersion="3.6.0">
-  <bpmn:process id="Process_07m7nhf" isExecutable="true">
+<bpmn:definitions xmlns:bpmn="http://www.omg.org/spec/BPMN/20100524/MODEL" xmlns:bpmndi="http://www.omg.org/spec/BPMN/20100524/DI" xmlns:dc="http://www.omg.org/spec/DD/20100524/DC" xmlns:camunda="http://camunda.org/schema/1.0/bpmn" xmlns:di="http://www.omg.org/spec/DD/20100524/DI" id="Definitions_0e0r9n9" targetNamespace="http://bpmn.io/schema/bpmn" exporter="Camunda Modeler" exporterVersion="4.1.1">
+  <bpmn:process id="email-trials" name="email-trials" isExecutable="true">
     <bpmn:startEvent id="StartEvent_1">
       <bpmn:outgoing>SequenceFlow_16wx5hd</bpmn:outgoing>
     </bpmn:startEvent>
@@ -45,29 +45,29 @@
     </bpmn:serviceTask>
   </bpmn:process>
   <bpmndi:BPMNDiagram id="BPMNDiagram_1">
-    <bpmndi:BPMNPlane id="BPMNPlane_1" bpmnElement="Process_07m7nhf">
-      <bpmndi:BPMNShape id="_BPMNShape_StartEvent_2" bpmnElement="StartEvent_1">
-        <dc:Bounds x="179" y="102" width="36" height="36" />
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNEdge id="SequenceFlow_16wx5hd_di" bpmnElement="SequenceFlow_16wx5hd">
-        <di:waypoint x="215" y="120" />
-        <di:waypoint x="270" y="120" />
-      </bpmndi:BPMNEdge>
-      <bpmndi:BPMNEdge id="SequenceFlow_08fcw9z_di" bpmnElement="SequenceFlow_08fcw9z">
-        <di:waypoint x="370" y="120" />
-        <di:waypoint x="430" y="120" />
+    <bpmndi:BPMNPlane id="BPMNPlane_1" bpmnElement="email-trials">
+      <bpmndi:BPMNEdge id="SequenceFlow_1bjqwqh_di" bpmnElement="SequenceFlow_1bjqwqh">
+        <di:waypoint x="690" y="120" />
+        <di:waypoint x="752" y="120" />
       </bpmndi:BPMNEdge>
       <bpmndi:BPMNEdge id="SequenceFlow_1g30hku_di" bpmnElement="SequenceFlow_1g30hku">
         <di:waypoint x="530" y="120" />
         <di:waypoint x="590" y="120" />
       </bpmndi:BPMNEdge>
+      <bpmndi:BPMNEdge id="SequenceFlow_08fcw9z_di" bpmnElement="SequenceFlow_08fcw9z">
+        <di:waypoint x="370" y="120" />
+        <di:waypoint x="430" y="120" />
+      </bpmndi:BPMNEdge>
+      <bpmndi:BPMNEdge id="SequenceFlow_16wx5hd_di" bpmnElement="SequenceFlow_16wx5hd">
+        <di:waypoint x="215" y="120" />
+        <di:waypoint x="270" y="120" />
+      </bpmndi:BPMNEdge>
+      <bpmndi:BPMNShape id="_BPMNShape_StartEvent_2" bpmnElement="StartEvent_1">
+        <dc:Bounds x="179" y="102" width="36" height="36" />
+      </bpmndi:BPMNShape>
       <bpmndi:BPMNShape id="EndEvent_1tmu3nh_di" bpmnElement="EndEvent_1tmu3nh">
         <dc:Bounds x="752" y="102" width="36" height="36" />
       </bpmndi:BPMNShape>
-      <bpmndi:BPMNEdge id="SequenceFlow_1bjqwqh_di" bpmnElement="SequenceFlow_1bjqwqh">
-        <di:waypoint x="690" y="120" />
-        <di:waypoint x="752" y="120" />
-      </bpmndi:BPMNEdge>
       <bpmndi:BPMNShape id="UserTask_1fd54vi_di" bpmnElement="firstTask">
         <dc:Bounds x="270" y="80" width="100" height="80" />
       </bpmndi:BPMNShape>
