+++ src/test/resources/camunda.cfg.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<beans xmlns="http://www.springframework.org/schema/beans" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xsi:schemaLocation="http://www.springframework.org/schema/beans   http://www.springframework.org/schema/beans/spring-beans.xsd">
+
+    <bean id="processEngineConfiguration" class="org.camunda.bpm.extension.process_test_coverage.junit.rules.ProcessCoverageInMemProcessEngineConfiguration">
+
+        <property name="history" value="full" />
+        <property name="databaseSchemaUpdate" value="true" />
+        <property name="jobExecutorActivate" value="false" />
+        <property name="expressionManager">
+            <bean class="org.camunda.bpm.engine.test.mock.MockExpressionManager" />
+        </property>
+        <property name="idGenerator">
+            <bean class="org.camunda.bpm.engine.impl.persistence.StrongUuidGenerator"/>
+        </property>
+
+    </bean>
+</beans>
